\name{OutbreakDetector-package}
\alias{OutbreakDetector-package}
\alias{OutbreakDetector}
\docType{package}
\title{
\packageTitle{OutbreakDetector}
}
\description{
\packageDescription{OutbreakDetector}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OutbreakDetector}
\packageIndices{OutbreakDetector}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{OutbreakDetector}

Maintainer: \packageMaintainer{OutbreakDetector}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
